+++ Makefile
@@ -51,6 +51,7 @@ SRC_FILES	=	main.c \
 				parser_info.c \
 				parser_map.c \
 				player_hook.c \
+				raycasting_find.c \
 				raycasting.c \
 				utils_mlx.c \
 				utils_vec.c \
