+++ src/init_player.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init_player.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: alefranc <alefranc@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/06 16:53:50 by alefranc          #+#    #+#             */
+/*   Updated: 2022/09/06 16:58:19 by alefranc         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "cub3d.h"
 
 static char	find_player_position(char **map, int *x, int *y)
@@ -60,16 +72,10 @@ static void	init_player_dir(char **map, t_vec *dir)
 	}
 }
 
-void	player_plane(t_all *all)
-{
-	all->player.plane.x = -FOVRATIO * all->player.dir.y;
-	all->player.plane.y = FOVRATIO * all->player.dir.x;
-}
-
 static void	remove_player_map_position(t_all *all)
 {
-	size_t x;
-	size_t y;
+	size_t	x;
+	size_t	y;
 
 	y = 0;
 	while (y < ft_strtabsize(all->map))
@@ -77,7 +83,8 @@ static void	remove_player_map_position(t_all *all)
 		x = 0;
 		while (x < ft_strlen(all->map[y]))
 		{
-			if (all->map[y][x] == 'N' ||all->map[y][x] == 'S' ||all->map[y][x] == 'W' ||all->map[y][x] == 'E' )
+			if (all->map[y][x] == 'N' || all->map[y][x] == 'S'
+				|| all->map[y][x] == 'W' || all->map[y][x] == 'E' )
 				all->map[y][x] = '0';
 			x++;
 		}
@@ -85,7 +92,6 @@ static void	remove_player_map_position(t_all *all)
 	}
 }
 
-// 	init_player(&all->player, all->map);
 void	init_player(t_all *all)
 {
 	init_player_pos(all->map, &all->player.pos);
@@ -93,58 +99,3 @@ void	init_player(t_all *all)
 	player_plane(all);
 	remove_player_map_position(all);
 }
-//
-// void	print_map(char **map)
-// {
-// 	int i;
-//
-// 	i = 0;
-// 	while (map[i] != NULL)
-// 	{
-// 		printf("%s\n", map[i]);
-// 		i++;
-// 	}
-// }
-
-// int	main(int argc, char **argv)
-// {
-// 	// t_all	all;
-// 	// char	*valid_map2[] =
-// 	// {   "       11111111111111111111111111",
-// 	// 	"       1111111111111111111111111",
-// 	// 	"       11111W00001110000111111111",
-// 	// 	"111111100000000000000001         ",
-// 	// 	"111111111011000001110000111111111",
-// 	// 	"100000000011000001110111111111111",
-// 	// 	"11110111111111011100000010001    ",
-// 	// 	"11110111111111011101010010001    ",
-// 	// 	"11111000110101011100000010001    ",
-// 	// 	"     1000000000000000110000011111",
-// 	// 	"11111000000000001101010010001",
-// 	// 	"110000011101010111110111100111",
-// 	// 	"11110111 1110101 101111010001",
-// 	// 	"11111111 1111111 111111111111", NULL};
-// 	//
-// 	// t_player player;
-// 	// all.map = valid_map2;
-//
-// 	t_all	*all;
-//
-// 	all = init_all();
-// 	if (all == NULL)
-// 		return (1);
-// 	if (parser(argc, argv, all) != 0)
-// 		return (destroy_all(all), 1);
-//
-// 	// all.player = player;
-// 	print_map(all->map);
-// 	remove_player_map_position(all);
-// 	print_map(all->map);
-// 	destroy_all(all);
-// 	// printf("pos.x = %f\n", player.pos.x);
-// 	// printf("pos.y = %f\n", player.pos.y);
-// 	// printf("dir.x = %f\n", player.dir.x);
-// 	// printf("dir.y = %f\n", player.dir.y);
-// 	// printf("plane.x = %f\n", player.plane.x);
-// 	// printf("plane.y = %f\n", player.plane.y);
-// }
