+++ src/minimap_display.c
@@ -6,33 +6,83 @@
 /*   By: alefranc <alefranc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/26 17:07:25 by alefranc          #+#    #+#             */
-/*   Updated: 2022/09/06 16:20:46 by lmarecha         ###   ########.fr       */
+/*   Updated: 2022/09/06 16:49:44 by alefranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void	draw_grid(t_all *all, t_data *img)
+static void	draw_line(t_data *img, t_point start, t_point end, int i)
 {
-	size_t	x;
-	size_t	y;
-	size_t	xmax;
-	size_t	ymax;
+	float	x;
+	float	y;
+	float	dx;
+	float	dy;
+	float	step;
 
-	y = 0;
-	xmax = ft_strlen(all->map[y]);
-	ymax = ft_strtabsize(all->map);
-	while (y < ymax * TILE)
+	dx = (end.x - start.x);
+	dy = (end.y - start.y);
+	if (fabs(dx) >= fabs(dy))
+		step = fabs(dx);
+	else
+		step = fabs(dy);
+	dx = dx / step;
+	dy = dy / step;
+	x = start.x;
+	y = start.y;
+	i = 1;
+	while (i <= step)
 	{
-		x = 0;
-		while (x < xmax * TILE)
-		{
-			if (y % TILE == 0 || x % TILE == 0)
-			my_mlx_pixel_put(img, x, y, 0x00000000);
-			x++;
-		}
-		y++;
+		my_mlx_pixel_put(img, (int)x, (int)y, 0x000000FF);
+		x = x + dx;
+		y = y + dy;
+		i = i + 1;
 	}
+}
+
+static void	draw_player_box(t_data *img, t_point start, int color)
+{
+	int	i;
+	int	j;
+
+	i = -2;
+	while (i < 3)
+	{
+		j = -2;
+		while (j < 3)
+		{
+			my_mlx_pixel_put(img, start.x + i, start.y + j, color);
+			j++;
+		}
+		i++;
+	}
+}
+
+void	display_player(t_all *all, t_data *img)
+{
+	t_point	start;
+	t_point	end;
+
+	start.x = (int)(all->player.pos.x * TILE);
+	start.y = (int)(all->player.pos.y * TILE);
+	end.x = (int)((all->player.pos.x + all->player.dir.x) * TILE);
+	end.y = (int)((all->player.pos.y + all->player.dir.y) * TILE);
+	draw_line(img, start, end, 0);
+	start.x = (int)(all->player.pos.x * TILE);
+	start.y = (int)(all->player.pos.y * TILE);
+	end.x = (int)((all->player.pos.x
+				+ (all->player.dir.x + all->player.plane.x) / 2) * TILE);
+	end.y = (int)((all->player.pos.y
+				+ (all->player.dir.y + all->player.plane.y) / 2) * TILE);
+	draw_line(img, start, end, 0);
+	start.x = (int)(all->player.pos.x * TILE);
+	start.y = (int)(all->player.pos.y * TILE);
+	end.x = (int)((all->player.pos.x
+				+ (all->player.dir.x - all->player.plane.x) / 2) * TILE);
+	end.y = (int)((all->player.pos.y
+				+ (all->player.dir.y - all->player.plane.y) / 2) * TILE);
+	draw_line(img, start, end, 0);
+	draw_player_box(img, start, 0x00FF4500);
 }
 
 void	draw_minimap(t_all *all, t_data *img)
@@ -59,22 +109,5 @@ void	draw_minimap(t_all *all, t_data *img)
 		}
 		y++;
 	}
-	// draw_grid(all, img);
 	display_player(all, img);
-}
-
-
-int	display_minimap(t_all *all, int offsetx, int offsety)
-{
-	t_data	img;
-
-	img.img = mlx_new_image(all->mlx, ft_strlen(all->map[0]) * TILE,
-			ft_strtabsize(all->map) * TILE);
-	img.addr = (unsigned int*)mlx_get_data_addr(img.img, &img.bits_per_pixel, &img.line_length,
-			&img.endian);
-	draw_minimap(all, &img);
-	draw_grid(all, &img);
-	mlx_put_image_to_window(all->mlx, all->win, img.img, offsetx, offsety);
-	mlx_destroy_image(all->mlx, img.img);
-	return (0);
 }
