+++ src/utils_vec.c
@@ -6,7 +6,7 @@
 /*   By: alefranc <alefranc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 14:17:12 by alefranc          #+#    #+#             */
-/*   Updated: 2022/08/31 14:19:57 by alefranc         ###   ########.fr       */
+/*   Updated: 2022/09/06 16:57:34 by alefranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,4 +15,10 @@
 double	norm(t_vec v)
 {
 	return (sqrt(v.x * v.x + v.y * v.y));
+}
+
+void	player_plane(t_all *all)
+{
+	all->player.plane.x = -FOVRATIO * all->player.dir.y;
+	all->player.plane.y = FOVRATIO * all->player.dir.x;
 }
